requirements.txt
ollama
pyexiftool
Pillow>=10.0
